chore(EMS-1721): split up account sign in conditions, fix SonarCloud code smells
