+++ src/api/.keystone/config.js
@@ -2116,9 +2116,6 @@ var sendEmailConfirmEmailAddressMutation = async (root, variables, context) => {
 };
 var send_email_confirm_email_address_default2 = sendEmailConfirmEmailAddressMutation;
 
-// custom-resolvers/mutations/account-sign-in/index.ts
-var import_date_fns5 = require("date-fns");
-
 // helpers/get-password-hash/index.ts
 var import_crypto3 = __toESM(require("crypto"));
 var { ENCRYPTION: ENCRYPTION3 } = ACCOUNT2;
@@ -2147,68 +2144,6 @@ var isValidAccountPassword = (password2, salt, hash) => {
   return false;
 };
 var is_valid_account_password_default = isValidAccountPassword;
-
-// helpers/generate-otp/index.ts
-var import_crypto4 = __toESM(require("crypto"));
-var import_otplib = require("otplib");
-var { ENCRYPTION: ENCRYPTION4, OTP } = ACCOUNT2;
-var {
-  RANDOM_BYTES_SIZE: RANDOM_BYTES_SIZE2,
-  STRING_TYPE: STRING_TYPE4,
-  PBKDF2: { ITERATIONS: ITERATIONS4, DIGEST_ALGORITHM: DIGEST_ALGORITHM4 },
-  OTP: {
-    PBKDF2: { KEY_LENGTH: KEY_LENGTH4 }
-  }
-} = ENCRYPTION4;
-var generateOtp = () => {
-  try {
-    console.info("Generating OTP");
-    const salt = import_crypto4.default.randomBytes(RANDOM_BYTES_SIZE2).toString(STRING_TYPE4);
-    import_otplib.authenticator.options = { digits: OTP.DIGITS };
-    const securityCode = import_otplib.authenticator.generate(salt);
-    const hash = import_crypto4.default.pbkdf2Sync(securityCode, salt, ITERATIONS4, KEY_LENGTH4, DIGEST_ALGORITHM4).toString(STRING_TYPE4);
-    const expiry = OTP.VERIFICATION_EXPIRY();
-    return {
-      securityCode,
-      salt,
-      hash,
-      expiry
-    };
-  } catch (err) {
-    console.error(err);
-    throw new Error(`Error generating OTP ${err}`);
-  }
-};
-var generate = {
-  otp: generateOtp
-};
-var generate_otp_default = generate;
-
-// helpers/generate-otp-and-update-account/index.ts
-var generateOTPAndUpdateAccount = async (context, accountId) => {
-  try {
-    console.info("Adding OTP to an account");
-    const otp = generate_otp_default.otp();
-    const { securityCode, salt, hash, expiry } = otp;
-    const accountUpdate = {
-      otpSalt: salt,
-      otpHash: hash,
-      otpExpiry: expiry
-    };
-    await context.db.Account.updateOne({
-      where: { id: accountId },
-      data: accountUpdate
-    });
-    return {
-      success: true,
-      securityCode
-    };
-  } catch (err) {
-    console.error(err);
-    throw new Error(`Adding OTP to an account ${err}`);
-  }
-};
-var generate_otp_and_update_account_default = generateOTPAndUpdateAccount;
 
 // helpers/create-authentication-retry-entry/index.ts
 var createAuthenticationRetryEntry = async (context, accountId) => {
@@ -2287,8 +2222,118 @@ var blockAccount = async (context, accountId) => {
 };
 var block_account_default = blockAccount;
 
-// custom-resolvers/mutations/account-sign-in/index.ts
+// custom-resolvers/mutations/account-sign-in/account-checks/index.ts
+var import_date_fns5 = require("date-fns");
+
+// helpers/generate-otp/index.ts
+var import_crypto4 = __toESM(require("crypto"));
+var import_otplib = require("otplib");
+var { ENCRYPTION: ENCRYPTION4, OTP } = ACCOUNT2;
+var {
+  RANDOM_BYTES_SIZE: RANDOM_BYTES_SIZE2,
+  STRING_TYPE: STRING_TYPE4,
+  PBKDF2: { ITERATIONS: ITERATIONS4, DIGEST_ALGORITHM: DIGEST_ALGORITHM4 },
+  OTP: {
+    PBKDF2: { KEY_LENGTH: KEY_LENGTH4 }
+  }
+} = ENCRYPTION4;
+var generateOtp = () => {
+  try {
+    console.info("Generating OTP");
+    const salt = import_crypto4.default.randomBytes(RANDOM_BYTES_SIZE2).toString(STRING_TYPE4);
+    import_otplib.authenticator.options = { digits: OTP.DIGITS };
+    const securityCode = import_otplib.authenticator.generate(salt);
+    const hash = import_crypto4.default.pbkdf2Sync(securityCode, salt, ITERATIONS4, KEY_LENGTH4, DIGEST_ALGORITHM4).toString(STRING_TYPE4);
+    const expiry = OTP.VERIFICATION_EXPIRY();
+    return {
+      securityCode,
+      salt,
+      hash,
+      expiry
+    };
+  } catch (err) {
+    console.error(err);
+    throw new Error(`Error generating OTP ${err}`);
+  }
+};
+var generate = {
+  otp: generateOtp
+};
+var generate_otp_default = generate;
+
+// helpers/generate-otp-and-update-account/index.ts
+var generateOTPAndUpdateAccount = async (context, accountId) => {
+  try {
+    console.info("Adding OTP to an account");
+    const otp = generate_otp_default.otp();
+    const { securityCode, salt, hash, expiry } = otp;
+    const accountUpdate = {
+      otpSalt: salt,
+      otpHash: hash,
+      otpExpiry: expiry
+    };
+    await context.db.Account.updateOne({
+      where: { id: accountId },
+      data: accountUpdate
+    });
+    return {
+      success: true,
+      securityCode
+    };
+  } catch (err) {
+    console.error(err);
+    throw new Error(`Adding OTP to an account ${err}`);
+  }
+};
+var generate_otp_and_update_account_default = generateOTPAndUpdateAccount;
+
+// custom-resolvers/mutations/account-sign-in/account-checks/index.ts
 var { EMAIL: EMAIL2 } = ACCOUNT2;
+var accountChecks = async (context, account, urlOrigin) => {
+  const { id: accountId, email } = account;
+  if (!account.isVerified) {
+    console.info("Unable to sign in account - account has not been verified yet");
+    const now = /* @__PURE__ */ new Date();
+    const verificationHasExpired = (0, import_date_fns5.isAfter)(now, account.verificationExpiry);
+    if (account.verificationHash && !verificationHasExpired) {
+      console.info("Account has an unexpired verification token - resetting verification expiry");
+      const accountUpdate = {
+        verificationExpiry: EMAIL2.VERIFICATION_EXPIRY()
+      };
+      await context.db.Account.updateOne({
+        where: { id: accountId },
+        data: accountUpdate
+      });
+      console.info("Account has an unexpired verification token - sending verification email");
+      const emailResponse2 = await send_email_confirm_email_address_default.send(context, urlOrigin, accountId);
+      if (emailResponse2.success) {
+        return {
+          success: false,
+          resentVerificationEmail: true,
+          accountId
+        };
+      }
+      return { success: false, accountId };
+    }
+    console.info("Unable to sign in account - account has not been verified");
+    return { success: false };
+  }
+  const { securityCode } = await generate_otp_and_update_account_default(context, accountId);
+  const name = get_full_name_string_default(account);
+  const emailResponse = await emails_default.securityCodeEmail(email, name, securityCode);
+  if (emailResponse.success) {
+    return {
+      ...emailResponse,
+      accountId
+    };
+  }
+  return {
+    success: false
+  };
+};
+var account_checks_default = accountChecks;
+
+// custom-resolvers/mutations/account-sign-in/index.ts
 var accountSignIn = async (root, variables, context) => {
   try {
     console.info("Signing in account");
@@ -2322,45 +2367,7 @@ var accountSignIn = async (root, variables, context) => {
       return { success: false };
     }
     if (is_valid_account_password_default(password2, account.salt, account.hash)) {
-      if (!account.isVerified) {
-        console.info("Unable to sign in account - account has not been verified yet");
-        const now = /* @__PURE__ */ new Date();
-        const verificationHasExpired = (0, import_date_fns5.isAfter)(now, account.verificationExpiry);
-        if (account.verificationHash && !verificationHasExpired) {
-          console.info("Account has an unexpired verification token - resetting verification expiry");
-          const accountUpdate = {
-            verificationExpiry: EMAIL2.VERIFICATION_EXPIRY()
-          };
-          await context.db.Account.updateOne({
-            where: { id: accountId },
-            data: accountUpdate
-          });
-          console.info("Account has an unexpired verification token - sending verification email");
-          const emailResponse2 = await send_email_confirm_email_address_default.send(context, urlOrigin, accountId);
-          if (emailResponse2.success) {
-            return {
-              success: false,
-              resentVerificationEmail: true,
-              accountId
-            };
-          }
-          return { success: false, accountId };
-        }
-        console.info("Unable to sign in account - account has not been verified");
-        return { success: false };
-      }
-      const { securityCode } = await generate_otp_and_update_account_default(context, accountId);
-      const name = get_full_name_string_default(account);
-      const emailResponse = await emails_default.securityCodeEmail(email, name, securityCode);
-      if (emailResponse.success) {
-        return {
-          ...emailResponse,
-          accountId
-        };
-      }
-      return {
-        success: false
-      };
+      return account_checks_default(context, account, urlOrigin);
     }
     return { success: false };
   } catch (err) {
@@ -2888,7 +2895,7 @@ var updateCompanyAndCompanyAddress = async (root, variables, context) => {
         where: oldSicCodes
       });
     }
-    if (mappedSicCodes && mappedSicCodes.length) {
+    if (mappedSicCodes?.length) {
       await context.db.CompanySicCode.createMany({
         data: mappedSicCodes
       });
