+++ src/api/custom-resolvers/mutations/account-sign-in/index.ts
@@ -1,18 +1,12 @@
 import { Context } from '.keystone/types'; // eslint-disable-line
-import { isAfter } from 'date-fns';
-import { ACCOUNT, FIELD_IDS } from '../../../constants';
+import { FIELD_IDS } from '../../../constants';
 import getAccountByField from '../../../helpers/get-account-by-field';
-import confirmEmailAddressEmail from '../../../helpers/send-email-confirm-email-address';
 import isValidAccountPassword from '../../../helpers/is-valid-account-password';
-import generateOTPAndUpdateAccount from '../../../helpers/generate-otp-and-update-account';
-import getFullNameString from '../../../helpers/get-full-name-string';
 import createAuthenticationRetryEntry from '../../../helpers/create-authentication-retry-entry';
 import shouldBlockAccount from '../../../helpers/should-block-account';
 import blockAccount from '../../../helpers/block-account';
-import sendEmail from '../../../emails';
+import accountChecks from './account-checks';
 import { Account, AccountSignInVariables, AccountSignInResponse } from '../../../types';
-
-const { EMAIL } = ACCOUNT;
 
 /**
  * accountSignIn
@@ -34,7 +28,7 @@ const accountSignIn = async (root: any, variables: AccountSignInVariables, conte
     const { urlOrigin, email, password } = variables;
 
     // Get the account the email is associated with.
-    const accountData = await getAccountByField(context, FIELD_IDS.INSURANCE.ACCOUNT.EMAIL, email);
+    const accountData = await getAccountByField(context, FIELD_IDS.INSURANCE.ACCOUNT.EMAIL, email) as Account;
 
     if (!accountData) {
       console.info('Unable to validate account - no account found');
@@ -42,7 +36,7 @@ const accountSignIn = async (root: any, variables: AccountSignInVariables, conte
       return { success: false };
     }
 
-    const account = accountData as Account;
+    const account = accountData;
 
     const { id: accountId } = account;
 
@@ -60,7 +54,7 @@ const accountSignIn = async (root: any, variables: AccountSignInVariables, conte
      * Check if the account is blocked
      * If so, return isBlocked=false
      */
-    const { isBlocked } = account as Account;
+    const { isBlocked } = account;
 
     if (isBlocked) {
       console.info('Unable to sign in account - account is already blocked');
@@ -90,7 +84,7 @@ const accountSignIn = async (root: any, variables: AccountSignInVariables, conte
 
     /**
      * Account is found and verified. We can therefore:
-     * 1) Check if the password is matches what is encrypted in the database.
+     * 1) Check if the password matches what is encrypted in the database.
      * 2) If the password is valid:
      *   - If the account is unverified, but has a valid has/token, send verification email.
      *   - If the account is verified, generate an OTP/security code and send via email.
@@ -99,75 +93,7 @@ const accountSignIn = async (root: any, variables: AccountSignInVariables, conte
      *   - The email was not sent.
      */
     if (isValidAccountPassword(password, account.salt, account.hash)) {
-      /**
-       * Check if the account:
-       * 1) Is unverified.
-       * 2) Has a verification hash/token.
-       * 3) Has a verification hash/token that has not expired.
-       *
-       * This means that the user has not verified their email address.
-       * We can therefore:
-       * 1) Reset the account's verification expiry.
-       * 2) Re-send the verification link email that was sent during account creation.
-       */
-      if (!account.isVerified) {
-        console.info('Unable to sign in account - account has not been verified yet');
-
-        const now = new Date();
-
-        const verificationHasExpired = isAfter(now, account.verificationExpiry);
-
-        if (account.verificationHash && !verificationHasExpired) {
-          console.info('Account has an unexpired verification token - resetting verification expiry');
-
-          const accountUpdate = {
-            verificationExpiry: EMAIL.VERIFICATION_EXPIRY(),
-          };
-
-          (await context.db.Account.updateOne({
-            where: { id: accountId },
-            data: accountUpdate,
-          })) as Account;
-
-          console.info('Account has an unexpired verification token - sending verification email');
-
-          const emailResponse = await confirmEmailAddressEmail.send(context, urlOrigin, accountId);
-
-          if (emailResponse.success) {
-            return {
-              success: false,
-              resentVerificationEmail: true,
-              accountId,
-            };
-          }
-
-          return { success: false, accountId };
-        }
-
-        // reject
-        console.info('Unable to sign in account - account has not been verified');
-
-        return { success: false };
-      }
-
-      // generate OTP and update the account
-      const { securityCode } = await generateOTPAndUpdateAccount(context, accountId);
-
-      // send "security code" email.
-      const name = getFullNameString(account);
-
-      const emailResponse = await sendEmail.securityCodeEmail(email, name, securityCode);
-
-      if (emailResponse.success) {
-        return {
-          ...emailResponse,
-          accountId,
-        };
-      }
-
-      return {
-        success: false,
-      };
+      return accountChecks(context, account, urlOrigin);
     }
 
     // invalid credentials.
